WalDataStore

+++ src/main/java/org/hillhouse/searchdb/interfaces/dao/IDDao.java
@@ -1,5 +1,7 @@
 package org.hillhouse.searchdb.interfaces.dao;
 
+import java.util.concurrent.atomic.AtomicInteger;
+
 public interface IDDao {
-    long getNextID();
+    AtomicInteger getNextID();
 }
